Rename types.py to protocol.py to not shadow built-in types
