+++ src/mcbootflash/__init__.py
@@ -15,7 +15,7 @@ from .flash import (
     self_verify,
     write_flash,
 )
-from .types import BootAttrs, Chunk, Command
+from .protocol import BootAttrs, Chunk, Command
 from .util import chunked
 
 __all__ = [
